+++ openapi-generator-annotations/lib/src/openapi_generator_annotations_base.dart
@@ -122,6 +122,21 @@ class AdditionalProperties {
   /// Flutter wrapper to use (none|flutterw|fvm)
   final Wrapper wrapper;
 
+  /// Set to true for generators with better support for discriminators.
+  /// (Python, Java, Go, PowerShell, C#have this enabled by default).
+  ///
+  /// true
+  /// The mapping in the discriminator includes descendent schemas that allOf
+  /// inherit from self and the discriminator mapping schemas in the OAS document.
+  ///
+  /// false
+  /// The mapping in the discriminator includes any descendent schemas that allOf
+  /// inherit from self, any oneOf schemas, any anyOf schemas, any x-discriminator-values,
+  /// and the discriminator mapping schemas in the OAS document AND Codegen validates
+  /// that oneOf and anyOf schemas contain the required discriminator and throws
+  /// an error if the discriminator is missing.
+  final bool legacyDiscriminatorBehavior;
+
   const AdditionalProperties(
       {this.allowUnicodeIdentifiers = false,
       this.ensureUniqueParams = true,
@@ -131,6 +146,7 @@ class AdditionalProperties {
       this.pubAuthorEmail,
       this.pubDescription,
       this.pubHomepage,
+      this.legacyDiscriminatorBehavior = true,
       this.pubName,
       this.pubVersion,
       this.sortModelPropertiesByRequiredFlag = true,
@@ -219,9 +235,52 @@ class DioProperties extends AdditionalProperties {
             useEnumExtension: useEnumExtension);
 }
 
-enum DioDateLibrary { core, timemachine }
+enum DioDateLibrary {
+  /// Dart core library (DateTime)
+  core,
+
+  /// Time Machine is date and time library for Flutter, Web, and Server with
+  /// support for timezones, calendars, cultures, formatting and parsing.
+  timemachine
+}
 enum SerializationFormat { JSON, PROTO }
 
 /// The name of the generator to use
-enum Generator { DART, DART_DIO, DART2_API, DART_JAGUAR }
-enum Wrapper {fvm, flutterw, none}
+enum Generator {
+  @Deprecated('Use Generator.dart instead')
+  DART,
+
+  @Deprecated('Use Generator.dio instead')
+  DART_DIO,
+
+  @Deprecated('Use Generator.dioAlt instead')
+  DART2_API,
+
+  @Deprecated('Use Generator.jaguar instead')
+  DART_JAGUAR,
+
+  /// This generator uses the default http package that comes with dart
+  /// corresponds to dart
+  dart,
+
+  /// This generator uses the dio package. Source gen is required after generating code with this generator
+  /// corresponds to dart-dio
+  ///
+  /// A powerful Http client for Dart, which supports Interceptors, Global configuration,
+  /// FormData, Request Cancellation, File downloading, Timeout etc
+  /// https://pub.flutter-io.cn/packages/dio
+  dio,
+
+  /// This uses the generator provided by bluetrainsoftware which internally uses the dio packaget
+  ///
+  /// You can read more about it here https://github.com/dart-ogurets/dart-openapi-maven
+  dioAlt,
+
+  /// This generates code based on the jaguar package Source gen is required
+  /// after generating code with this generator
+  /// corresponds to dart-jaguar
+  ///
+  /// An Http Api generator inspired by Retrofit for Dart
+  jaguar,
+}
+enum Wrapper { fvm, flutterw, none }